fix concurrency race when updating resources
You cannot release a lock and then call a function that acquires the
lock and assume that nothing happens in between. This rendered the
sorting ineffectively if you were unlucky.

This holds the lock the whole time and splits the code into an
"implementation" part, that assumes the lock was acquired and a public
method that does not make this assumption and acquires the lock itself
(and then also calls the lock-less "implementation").