TUI: add initial version
The fancier TUI can be started via "-t fancy".

It consists of two modes:
- Overview: This presents a live overview about the all the DRBD
resources. If resources get added, the will show up. If a "movement key"
is pressed (arrows, page keys, vim-style j/k), the output is frozen and
the user can apply actions to resources (e.g., switching a resource to
primary). These commands act on the selected resource. This is then
menu-based per key-stroke to guide new(er) users. For example you first
press 'r' for role, and then 's' for secondary. Experienced users can
press "rs" quickly. So far the actions are not actully executed, but the
foundations are there.

- Detailed View: If a resource is selected, the user can press enter to
view the resource in more detail. This the part where we want to make
use of all the fance stuff termui provides. So far this enumerates the
volumes in the resourcse and shows some fake graphs. This is just a
starting point, but for example the number of widgets is already dynamic
(depending on the number of volumes in that case). So it should be
relatively easy to extend on that and show some real information.