filter view: be more strict when allowed, fix bug if nothing to filter
So far there was a div-by-zero when there was nothing to filter and the
user pressed a movement key like 'j'. Avoid that by checking if there is
actually anything to filter, and if not, don't pretend to.

Filtering is only allowed in the overview mode when the display is not
locked. It was to easy to press 'f' in locked mode (which is reserved
for commands). Change the sate machine accordingly.