Fix race condition when pruning old events
This fix prunes old events based on the timestamp of the poll
cycle 3 cycles before the current one. Each poll cycle's timestamp
is based on the earliest timestamp of the events of that poll cycle.
If no events are seen in a poll cycle, drbdtop falls back to using
an internal time reference.