utils: refactor/split getVersionInfo
This splits this higher level function into two, one for drbd-utils
version and one for DRBD kernel version. This has the disadvantage that
we call/parse drbdadm output twice. Calling it twice could be avoided
with a buffer. All in all having cleaner, reusable code is more
important. This code is currently called once.